# interface-specs/openapi/platform-api-v1.yaml
openapi: 3.0.3
info:
  title: 县域系统集成平台管理接口
  version: 0.1.0
  description: >-
    OpenAPI 是用机器可读文件说明 HTTP 接口的通用规范。本文件描述当前平台内部管理接口。
    业务接口只有在接口规范、已删除或遮盖敏感字段的样例、网络条件和负责人均确认后才能开发。
servers:
  - url: /api/v1
security:
  - hospitalIdentity: []
paths:
...
  /session/login:
    post:
      security: []
      operationId: login
      summary: 登录并建立服务端管理会话
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 已登录的当前用户
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUserResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { description: 登录名或密码错误 }
  /session/current:
    get:
      operationId: getCurrentUser
      summary: 查询当前登录的管理用户
      responses:
        '200':
          description: 当前用户，不包含密码或会话编号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUserResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /session/logout:
    post:
      operationId: logout
      summary: 退出并使当前管理会话失效
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200': { description: 当前会话已失效 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /session/password:
    post:
      operationId: changeCurrentPassword
      summary: 修改当前本地密码并使当前会话失效
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200': { description: 密码已修改，用户必须重新登录 }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /users/{userId}/roles:
    put:
      operationId: replaceUserRoles
      summary: 一次性替换用户角色，同时保留至少一个可用管理员
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceRoleIdsRequest' }
      responses:
        '200': { description: 已替换用户角色 }
        '403': { description: 'access:write permission and organization scope are required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
...
  /roles:
    get:
      operationId: listRoles
      summary: 查询平台角色和后端已登记权限
      responses:
        '200':
          description: 平台角色
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleListResponse' }
        '403': { description: 'access:read permission is required' }
    post:
      operationId: createRole
      summary: 创建非系统保护角色
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoleRequest' }
      responses:
        '201': { description: 已创建角色 }
        '403': { description: 'access:write permission is required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /roles/{roleId}:
    parameters:
      - name: roleId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    get:
      operationId: getRole
      summary: 查询平台角色
      responses:
        '200': { description: 角色详情 }
        '403': { description: 'access:read permission is required' }
    put:
      operationId: updateRole
      summary: 修改非系统保护角色
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoleRequest' }
      responses:
        '200': { description: 已修改角色 }
        '403': { description: 'access:write permission is required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
    delete:
      operationId: deleteRole
      summary: 删除未分配给用户的非系统保护角色
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteVersionRequest' }
      responses:
        '200': { description: 角色已删除 }
        '403': { description: 'access:write permission is required' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /roles/{roleId}/permissions:
    put:
      operationId: replaceRolePermissions
      summary: 替换非系统保护角色的已登记权限
      parameters:
        - name: roleId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplacePermissionCodesRequest' }
      responses:
        '200': { description: 已替换角色权限 }
        '403': { description: 'access:write permission is required' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
  /permissions:
    get:
      operationId: listRegisteredPermissions
      summary: 查询后端代码已登记的权限
      responses:
        '200': { description: 后端已登记的权限列表 }
        '403': { description: 'access:read permission is required' }
...
components:
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: 由 GET /session/csrf 返回，并与 XSRF-TOKEN Cookie 配套使用的令牌
      schema: { type: string }
  securitySchemes:
    hospitalIdentity:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >-
        Cookie marker for an authenticated management session. The first development
        phase uses a server-side local-management session and secure HttpOnly cookie;
        production identity must be reviewed by the hospital before activation.
  responses:
    InvalidRequest:
      description: 请求参数校验未通过
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: 必须先登录
    ResourceNotFound:
      description: 请求的平台数据不存在，或当前账号无权查看
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ResourceConflict:
      description: 唯一值重复、当前状态不允许操作，或数据已被他人修改
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    LoginRequest:
      type: object
      required: [loginName, password]
      properties:
        loginName: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 128, format: password, writeOnly: true }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1, maxLength: 128, format: password, writeOnly: true }
        newPassword: { type: string, minLength: 12, maxLength: 128, format: password, writeOnly: true }
    CurrentUserResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: SUCCESS }
        message: { type: string }
        data: { $ref: '#/components/schemas/CurrentUser' }
    CurrentUser:
      type: object
      required: [userId, loginName, displayName, primaryOrganizationId, organizationCode, mustChangePassword, permissions, organizationCodes]
      properties:
        userId: { type: integer, format: int64 }
        loginName: { type: string }
        displayName: { type: string }
        primaryOrganizationId: { type: integer, format: int64 }
        organizationCode: { type: string }
        mustChangePassword: { type: boolean }
        permissions:
          type: array
          items: { type: string }
        organizationCodes:
          type: array
          description: 明确列出的已启用机构；可访问上级机构不表示自动获得下级机构权限
          items: { type: string }
...
    ReplaceRoleIdsRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds: { type: array, uniqueItems: true, items: { type: integer, format: int64, minimum: 1 } }
...
    RoleListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/Role' } }
    Role:
      type: object
      required: [id, roleCode, roleName, enabled, systemManaged, createdAt, updatedAt, version, permissionCodes]
      properties:
        id: { type: integer, format: int64 }
        roleCode: { type: string }
        roleName: { type: string }
        enabled: { type: boolean }
        systemManaged: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: string, format: byte }
        permissionCodes: { type: array, items: { type: string } }
    CreateRoleRequest:
      type: object
      required: [roleCode, roleName]
      properties:
        roleCode: { type: string, minLength: 3, maxLength: 64, pattern: '^[A-Z][A-Z0-9_]+$' }
        roleName: { type: string, minLength: 1, maxLength: 100 }
    UpdateRoleRequest:
      type: object
      required: [roleName, enabled, version]
      properties:
        roleName: { type: string, minLength: 1, maxLength: 100 }
        enabled: { type: boolean }
        version: { type: string, format: byte }
    DeleteVersionRequest:
      type: object
      required: [version]
      properties:
        version: { type: string, format: byte, description: 由 8 字节 SQL Server 行版本号编码得到的 Base64 文本 }
    ReplacePermissionCodesRequest:
      type: object
      required: [permissionCodes]
      properties:
        permissionCodes: { type: array, uniqueItems: true, items: { type: string, minLength: 1, maxLength: 64 } }
...
    ApiError:
      type: object
      required: [code, message, timestamp]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }